feat: Add MageForge Toolbar functionality with menu controls, item management, and score animations
